moved test code
