+++ tests/components/gdacs/test_init.py
@@ -1,0 +1,26 @@
+"""Define tests for the GDACS general setup."""
+from asynctest import CoroutineMock, patch
+
+from homeassistant.components.gdacs import (
+    DOMAIN,
+    FEED,
+    async_setup_entry,
+    async_unload_entry,
+)
+
+
+async def test_component_unload_config_entry(hass, config_entry):
+    """Test that loading and unloading of a config entry works."""
+    config_entry.add_to_hass(hass)
+    with patch(
+        "aio_georss_gdacs.GdacsFeedManager.update", new_callable=CoroutineMock
+    ) as mock_feed_manager_update:
+        # Load config entry.
+        assert await async_setup_entry(hass, config_entry)
+        await hass.async_block_till_done()
+        assert mock_feed_manager_update.call_count == 1
+        assert hass.data[DOMAIN][FEED][config_entry.entry_id] is not None
+        # Unload config entry.
+        assert await async_unload_entry(hass, config_entry)
+        await hass.async_block_till_done()
+        assert hass.data[DOMAIN][FEED].get(config_entry.entry_id) is None
